Create breakpoint.py
