+++ test/core/pipeline/test_pipeline_breakpoints_branch_joiner.py
@@ -13,6 +13,7 @@ from haystack.dataclasses import ChatMessage
 from haystack.utils.auth import Secret
 from haystack_experimental.core.errors import PipelineBreakpointException
 from haystack_experimental.core.pipeline.pipeline import Pipeline
+from haystack_experimental.core.pipeline.breakpoint import load_state
 from unittest.mock import patch
 
 class TestPipelineBreakpoints:
@@ -116,7 +117,7 @@ class TestPipelineBreakpoints:
             f_name = Path(full_path).name
             if str(f_name).startswith(component):
                 file_found = True
-                resume_state = Pipeline.load_state(full_path)
+                resume_state = load_state(full_path)
                 result = branch_joiner_pipeline.run(data, resume_state=resume_state)
                 assert result['validator']
                 break
